requirements.txt
monai==1.1.0
opencv-python==4.5.4.60
requests==2.28.1
scikit-image==0.19.3
scikit-learn==0.24.2
scipy==1.7.3
seaborn==0.12.1
timm==0.6.12
torch-summary==1.4.5
tqdm==4.64.1

